virtual-serial-communication-project/requirements.txt
click==8.1.3
colorama==0.4.6
mysql-connector-python==8.0.33
protobuf==3.20.3
pynput==1.7.6
PyQt5Designer==5.14.1
PyQt6==6.4.2

PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.1
pyqt6-tools==6.4.2.3.3
pyserial==3.5
python-dotenv==1.0.0
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
six==1.16.0
